moved connected clients to a seperate file
